chore(proof/tee): delete base-enclave crate, inline types into nitro-enclave
base-enclave's sole remaining types (PerChainConfig, BlockId, Genesis,
GenesisSystemConfig) had a single consumer (nitro-enclave). Inline them
into the nitro-enclave types module to eliminate the extra crate.

- Move config types and tests into nitro-enclave/src/types.rs
- Move dependencies (alloy-chains, alloy-eips, base-consensus-genesis)
  to nitro-enclave Cargo.toml
- Remove base-enclave workspace member and dependency entries
- Delete crates/proof/tee/core/ entirely
