--- crates/proof/tee/core/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "base-enclave"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# Alloy
-alloy-chains.workspace = true
-alloy-eips = { workspace = true, features = ["serde"] }
-alloy-primitives = { workspace = true, features = ["serde"] }
-
-# Consensus
-base-consensus-genesis = { workspace = true, features = ["serde"] }
-
-# Misc
-serde = { workspace = true, features = ["derive"] }
-
-[dev-dependencies]
-hex.workspace = true
-serde_json.workspace = true
-base-consensus-registry.workspace = true
