--- crates/proof/tee/core/src/lib.rs
@@ -1,4 +1,0 @@
-#![doc = include_str!("../README.md")]
-
-mod types;
-pub use types::{BlockId, Genesis, GenesisSystemConfig, PerChainConfig, RollupConfig};
