--- crates/proof/tee/core/src/types/mod.rs
@@ -1,4 +1,0 @@
-//! TEE-specific types for chain configuration.
-
-mod config;
-pub use config::{BlockId, Genesis, GenesisSystemConfig, PerChainConfig, RollupConfig};
